services/ml-anomaly-detector/requirements.txt
kafka-python
scikit-learn
joblib
numpy
PyYAML
